No heap allocations, no arrays, just one loop.
When he attack a even number of sheep in a star, he starts to
go to star i - 1. Since he already stole 1 from it, it now has an even
number of sheeps that will made him go back to the beggining and halts
his trek.
If a star had 1 sheep the first time he passed by it, if he starts the
"going back jorney" it will have 0, so he'll stole none. Because of it
I store each zero cases so it will be removed from the arithmetic operation
that gives us the number of stolen sheeps.
If a star has a even number a sheeps he stole 1 sheep from this star and
start to fall back, so he'll stole just 1 sheep from it. Because of it I
remove 1 in the formula.